+++ verifier/src/main/java/com/aws/jverify/verifier/compiler/simplifications/lambdas/LambdaCompiler.java
@@ -37,24 +37,18 @@ import com.sun.tools.javac.tree.TreeTranslator;
 import com.sun.tools.javac.code.Symbol.ClassSymbol;
 import com.sun.tools.javac.code.Symbol.DynamicMethodSymbol;
 import com.sun.tools.javac.code.Symbol.MethodSymbol;
-import com.sun.tools.javac.code.Symbol.TypeSymbol;
 import com.sun.tools.javac.code.Symbol.VarSymbol;
 import com.sun.tools.javac.code.Type.MethodType;
 import com.sun.tools.javac.code.Type.TypeVar;
-import com.aws.jverify.verifier.compiler.simplifications.lambdas.LambdaCompiler.LambdaAnalyzerPreprocessor.*;
+import com.aws.jverify.verifier.compiler.simplifications.lambdas.LambdaAnalyzerPreprocessor.*;
 //import com.sun.tools.javac.comp.Lower.BasicFreeVarCollector;
 //import com.sun.tools.javac.resources.CompilerProperties.Notes;
 import com.sun.tools.javac.jvm.*;
 import com.sun.tools.javac.util.*;
 import com.sun.source.tree.MemberReferenceTree.ReferenceMode;
 
-import java.util.EnumMap;
 import java.util.HashMap;
-import java.util.HashSet;
-import java.util.LinkedHashMap;
 import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
@@ -62,9 +56,7 @@ import static com.aws.jverify.verifier.compiler.simplifications.lambdas.LambdaCo
 import static com.sun.tools.javac.code.Flags.*;
 import static com.sun.tools.javac.code.Kinds.Kind.*;
 import static com.sun.tools.javac.code.TypeTag.*;
-import static com.sun.tools.javac.tree.JCTree.Tag.*;
 
-import javax.lang.model.element.ElementKind;
 import javax.lang.model.type.TypeKind;
 
 import com.sun.tools.javac.main.Option;
@@ -81,33 +73,33 @@ import com.sun.tools.javac.util.List;
 public class LambdaCompiler extends TreeTranslator {
 
     private Enter enter;
-    private Attr attr;
-    private JCDiagnostic.Factory diags;
-    private Log log;
+    public Attr attr;
+    public JCDiagnostic.Factory diags;
+    public Log log;
     private Lower lower;
-    private Names names;
-    private Symtab syms;
+    public Names names;
+    public Symtab syms;
     private Resolve rs;
     private Operators operators;
-    private TreeMaker make;
-    private Types types;
-    private TransTypes transTypes;
-    private Env<AttrContext> attrEnv;
+    public TreeMaker make;
+    public Types types;
+    public TransTypes transTypes;
+    public Env<AttrContext> attrEnv;
 
     /** the analyzer scanner */
     private LambdaAnalyzerPreprocessor analyzer;
 
     /** map from lambda trees to translation contexts */
-    private Map<JCTree, TranslationContext<?>> contextMap;
+    public Map<JCTree, TranslationContext<?>> contextMap;
 
     /** current translation context (visitor argument) */
-    private TranslationContext<?> context;
+    public TranslationContext<?> context;
 
     /** info about the current class being processed */
     private KlassInfo kInfo;
 
     /** dump statistics about lambda code generation */
-    private final boolean dumpLambdaToMethodStats;
+    public final boolean dumpLambdaToMethodStats;
 
     /** true if line or local variable debug info has been requested */
     private final boolean debugLinesOrVars;
@@ -119,7 +111,7 @@ public class LambdaCompiler extends TreeTranslator {
     private final boolean deduplicateLambdas;
 
     /** lambda proxy is a dynamic nestmate */
-    private final boolean nestmateLambdas;
+    public final boolean nestmateLambdas;
 
     /** Flag for alternate metafactories indicating the lambda object has multiple targets */
     public static final int FLAG_MARKERS = 1 << 1;
@@ -150,7 +142,7 @@ public class LambdaCompiler extends TreeTranslator {
         make = TreeMaker.instance(context);
         types = Types.instance(context);
         transTypes = TransTypes.instance(context);
-        analyzer = new LambdaAnalyzerPreprocessor();
+        analyzer = new LambdaAnalyzerPreprocessor(this);
         Options options = Options.instance(context);
         dumpLambdaToMethodStats = options.isSet("debug.dumpLambdaToMethodStats");
         attr = Attr.instance(context);
@@ -515,7 +507,7 @@ public class LambdaCompiler extends TreeTranslator {
     /**
      * Create new synthetic method with given flags, name, type, owner
      */
-    private MethodSymbol makePrivateSyntheticMethod(long flags, Name name, Type type, Symbol owner) {
+    public MethodSymbol makePrivateSyntheticMethod(long flags, Name name, Type type, Symbol owner) {
         return new MethodSymbol(flags | SYNTHETIC | PRIVATE, name, type, owner);
     }
 
@@ -530,7 +522,7 @@ public class LambdaCompiler extends TreeTranslator {
      * Set varargsElement field on a given tree (must be either a new class tree
      * or a method call tree)
      */
-    private void setVarargsIfNeeded(JCTree tree, Type varargsElement) {
+    public void setVarargsIfNeeded(JCTree tree, Type varargsElement) {
         if (varargsElement != null) {
             switch (tree.getTag()) {
                 case APPLY: ((JCMethodInvocation)tree).varargsElement = varargsElement; break;
@@ -547,7 +539,7 @@ public class LambdaCompiler extends TreeTranslator {
      * reference, as the bridged signature might require downcast to be compatible
      * with the generated signature.
      */
-    private List<JCExpression> convertArgs(Symbol meth, List<JCExpression> args, Type varargsElement) {
+    public List<JCExpression> convertArgs(Symbol meth, List<JCExpression> args, Type varargsElement) {
         Assert.check(meth.kind == MTH);
         List<Type> formals = meth.type.getParameterTypes();
         if (varargsElement != null) {
@@ -558,213 +550,6 @@ public class LambdaCompiler extends TreeTranslator {
 
     // </editor-fold>
 
-    /**
-     * Converts a method reference which cannot be used directly into a lambda
-     */
-    private class MemberReferenceToLambda {
-
-        private final JCMemberReference tree;
-        private final ReferenceTranslationContext localContext;
-        private final Symbol owner;
-        private final ListBuffer<JCExpression> args = new ListBuffer<>();
-        private final ListBuffer<JCVariableDecl> params = new ListBuffer<>();
-
-        private JCExpression receiverExpression = null;
-
-        MemberReferenceToLambda(JCMemberReference tree, ReferenceTranslationContext localContext, Symbol owner) {
-            this.tree = tree;
-            this.localContext = localContext;
-            this.owner = owner;
-        }
-
-        JCLambda lambda() {
-            int prevPos = make.pos;
-            try {
-                make.at(tree);
-
-                //body generation - this can be either a method call or a
-                //new instance creation expression, depending on the member reference kind
-                VarSymbol rcvr = addParametersReturnReceiver();
-                JCExpression expr = (tree.getMode() == ReferenceMode.INVOKE)
-                        ? expressionInvoke(rcvr)
-                        : expressionNew();
-
-                JCLambda slam = make.Lambda(params.toList(), expr);
-                slam.target = tree.target;
-                slam.type = tree.type;
-                slam.pos = tree.pos;
-                return slam;
-            } finally {
-                make.at(prevPos);
-            }
-        }
-
-        /**
-         * Generate the parameter list for the converted member reference.
-         *
-         * @return The receiver variable symbol, if any
-         */
-        VarSymbol addParametersReturnReceiver() {
-            Type samDesc = localContext.bridgedRefSig();
-            List<Type> samPTypes = samDesc.getParameterTypes();
-            List<Type> descPTypes = tree.getDescriptorType(types).getParameterTypes();
-
-            // Determine the receiver, if any
-            VarSymbol rcvr;
-            switch (tree.kind) {
-                case BOUND:
-                    // The receiver is explicit in the method reference
-                    rcvr = addParameter("rec$", tree.getQualifierExpression().type, false);
-                    receiverExpression = attr.makeNullCheck(tree.getQualifierExpression());
-                    break;
-                case UNBOUND:
-                    // The receiver is the first parameter, extract it and
-                    // adjust the SAM and unerased type lists accordingly
-                    rcvr = addParameter("rec$", samDesc.getParameterTypes().head, false);
-                    samPTypes = samPTypes.tail;
-                    descPTypes = descPTypes.tail;
-                    break;
-                default:
-                    rcvr = null;
-                    break;
-            }
-            List<Type> implPTypes = tree.sym.type.getParameterTypes();
-            int implSize = implPTypes.size();
-            int samSize = samPTypes.size();
-            // Last parameter to copy from referenced method, exclude final var args
-            int last = localContext.needsVarArgsConversion() ? implSize - 1 : implSize;
-
-            // Failsafe -- assure match-up
-            boolean checkForIntersection = tree.varargsElement != null || implSize == descPTypes.size();
-
-            // Use parameter types of the implementation method unless the unerased
-            // SAM parameter type is an intersection type, in that case use the
-            // erased SAM parameter type so that the supertype relationship
-            // the implementation method parameters is not obscured.
-            // Note: in this loop, the lists implPTypes, samPTypes, and descPTypes
-            // are used as pointers to the current parameter type information
-            // and are thus not usable afterwards.
-            for (int i = 0; implPTypes.nonEmpty() && i < last; ++i) {
-                // By default use the implementation method parameter type
-                Type parmType = implPTypes.head;
-                if (checkForIntersection) {
-                    if (descPTypes.head.getKind() == TypeKind.INTERSECTION) {
-                        parmType = samPTypes.head;
-                    }
-                    // If the unerased parameter type is a type variable whose
-                    // bound is an intersection (eg. <T extends A & B>) then
-                    // use the SAM parameter type
-                    if (descPTypes.head.getKind() == TypeKind.TYPEVAR) {
-                        TypeVar tv = (TypeVar) descPTypes.head;
-                        if (tv.getUpperBound().getKind() == TypeKind.INTERSECTION) {
-                            parmType = samPTypes.head;
-                        }
-                    }
-                }
-                addParameter("x$" + i, parmType, true);
-
-                // Advance to the next parameter
-                implPTypes = implPTypes.tail;
-                samPTypes = samPTypes.tail;
-                descPTypes = descPTypes.tail;
-            }
-            // Flatten out the var args
-            for (int i = last; i < samSize; ++i) {
-                addParameter("xva$" + i, tree.varargsElement, true);
-            }
-
-            return rcvr;
-        }
-
-        JCExpression getReceiverExpression() {
-            return receiverExpression;
-        }
-
-        private JCExpression makeReceiver(VarSymbol rcvr) {
-            if (rcvr == null) return null;
-            JCExpression rcvrExpr = makeIdent(rcvr);
-            boolean protAccess =
-                    isProtectedInSuperClassOfEnclosingClassInOtherPackage(tree.sym, owner);
-            Type rcvrType = tree.ownerAccessible && !protAccess ? tree.sym.enclClass().type
-                    : tree.expr.type;
-            if (rcvrType == syms.arrayClass.type) {
-                // Map the receiver type to the actually type, not just "array"
-                rcvrType = tree.getQualifierExpression().type;
-            }
-            if (!rcvr.type.tsym.isSubClass(rcvrType.tsym, types)) {
-                rcvrExpr = make.TypeCast(make.Type(rcvrType), rcvrExpr).setType(rcvrType);
-            }
-            return rcvrExpr;
-        }
-
-        /**
-         * determine the receiver of the method call - the receiver can
-         * be a type qualifier, the synthetic receiver parameter or 'super'.
-         */
-        private JCExpression expressionInvoke(VarSymbol rcvr) {
-            JCExpression qualifier =
-                    (rcvr != null) ?
-                            makeReceiver(rcvr) :
-                            tree.getQualifierExpression();
-
-            //create the qualifier expression
-            JCFieldAccess select = make.Select(qualifier, tree.sym);
-            select.sym = tree.sym;
-            select.type = tree.sym.erasure(types);
-
-            //create the method call expression
-            JCExpression apply = make.Apply(List.nil(), select,
-                            convertArgs(tree.sym, args.toList(), tree.varargsElement)).
-                    setType(tree.sym.erasure(types).getReturnType());
-
-            apply = transTypes.coerce(attrEnv, apply,
-                    localContext.tree.referentType.getReturnType());
-
-            setVarargsIfNeeded(apply, tree.varargsElement);
-            return apply;
-        }
-
-        /**
-         * Lambda body to use for a 'new'.
-         */
-        private JCExpression expressionNew() {
-            if (tree.kind == ReferenceKind.ARRAY_CTOR) {
-                //create the array creation expression
-                JCNewArray newArr = make.NewArray(
-                        make.Type(types.elemtype(tree.getQualifierExpression().type)),
-                        List.of(makeIdent(params.first().sym)),
-                        null);
-                newArr.type = tree.getQualifierExpression().type;
-                return newArr;
-            } else {
-                //create the instance creation expression
-                //note that method reference syntax does not allow an explicit
-                //enclosing class (so the enclosing class is null)
-                // but this may need to be patched up later with the proxy for the outer this
-                JCNewClass newClass = make.NewClass(null,
-                        List.nil(),
-                        make.Type(tree.getQualifierExpression().type),
-                        convertArgs(tree.sym, args.toList(), tree.varargsElement),
-                        null);
-                newClass.constructor = tree.sym;
-                newClass.constructorType = tree.sym.erasure(types);
-                newClass.type = tree.getQualifierExpression().type;
-                setVarargsIfNeeded(newClass, tree.varargsElement);
-                return newClass;
-            }
-        }
-
-        private VarSymbol addParameter(String name, Type p, boolean genArg) {
-            VarSymbol vsym = new VarSymbol(PARAMETER | SYNTHETIC, names.fromString(name), p, owner);
-            vsym.pos = tree.pos;
-            params.append(make.VarDef(vsym, null));
-            if (genArg) {
-                args.append(makeIdent(vsym));
-            }
-            return vsym;
-        }
-    }
-
     private MethodType typeToMethodType(Type mt) {
         Type type = mt;
         return new MethodType(type.getParameterTypes(),
@@ -774,1068 +559,8 @@ public class LambdaCompiler extends TreeTranslator {
     }
 
     // <editor-fold defaultstate="collapsed" desc="Lambda/reference analyzer">
-    /**
-     * This visitor collects information about translation of a lambda expression.
-     * More specifically, it keeps track of the enclosing contexts and captured locals
-     * accessed by the lambda being translated (as well as other useful info).
-     * It also translates away problems for LambdaToMethod.
-     */
-    class LambdaAnalyzerPreprocessor extends TreeTranslator {
 
-        /** the frame stack - used to reconstruct translation info about enclosing scopes */
-        private List<Frame> frameStack;
-
-        /**
-         * keep the count of lambda expression (used to generate unambiguous
-         * names)
-         */
-        private int lambdaCount = 0;
-
-        /**
-         * List of types undergoing construction via explicit constructor chaining.
-         */
-        private List<ClassSymbol> typesUnderConstruction;
-
-        /**
-         * keep the count of lambda expression defined in given context (used to
-         * generate unambiguous names for serializable lambdas)
-         */
-        private class SyntheticMethodNameCounter {
-            private Map<String, Integer> map = new HashMap<>();
-            int getIndex(StringBuilder buf) {
-                String temp = buf.toString();
-                Integer count = map.get(temp);
-                if (count == null) {
-                    count = 0;
-                }
-                ++count;
-                map.put(temp, count);
-                return count;
-            }
-        }
-        private SyntheticMethodNameCounter syntheticMethodNameCounts =
-                new SyntheticMethodNameCounter();
-
-        private Map<Symbol, JCClassDecl> localClassDefs;
-
-        /**
-         * maps for fake clinit symbols to be used as owners of lambda occurring in
-         * a static var init context
-         */
-        private Map<ClassSymbol, Symbol> clinits = new HashMap<>();
-
-        private JCClassDecl analyzeAndPreprocessClass(JCClassDecl tree) {
-            frameStack = List.nil();
-            typesUnderConstruction = List.nil();
-            localClassDefs = new HashMap<>();
-            return translate(tree);
-        }
-
-        @Override
-        public void visitApply(JCMethodInvocation tree) {
-            List<ClassSymbol> previousNascentTypes = typesUnderConstruction;
-            try {
-                Name methName = TreeInfo.name(tree.meth);
-                if (methName == names._this || methName == names._super) {
-                    typesUnderConstruction = typesUnderConstruction.prepend(currentClass());
-                }
-                super.visitApply(tree);
-            } finally {
-                typesUnderConstruction = previousNascentTypes;
-            }
-        }
-        // where
-        private ClassSymbol currentClass() {
-            for (Frame frame : frameStack) {
-                if (frame.tree.hasTag(JCTree.Tag.CLASSDEF)) {
-                    JCClassDecl cdef = (JCClassDecl) frame.tree;
-                    return cdef.sym;
-                }
-            }
-            return null;
-        }
-
-        @Override
-        public void visitBlock(JCBlock tree) {
-            List<Frame> prevStack = frameStack;
-            try {
-                if (frameStack.nonEmpty() && frameStack.head.tree.hasTag(CLASSDEF)) {
-                    frameStack = frameStack.prepend(new Frame(tree));
-                }
-                super.visitBlock(tree);
-            }
-            finally {
-                frameStack = prevStack;
-            }
-        }
-
-        @Override
-        public void visitClassDef(JCClassDecl tree) {
-            List<Frame> prevStack = frameStack;
-            int prevLambdaCount = lambdaCount;
-            SyntheticMethodNameCounter prevSyntheticMethodNameCounts =
-                    syntheticMethodNameCounts;
-            Map<ClassSymbol, Symbol> prevClinits = clinits;
-            DiagnosticSource prevSource = log.currentSource();
-            try {
-                log.useSource(tree.sym.sourcefile);
-                lambdaCount = 0;
-                syntheticMethodNameCounts = new SyntheticMethodNameCounter();
-                prevClinits = new HashMap<>();
-                if (tree.sym.owner.kind == MTH) {
-                    localClassDefs.put(tree.sym, tree);
-                }
-                if (directlyEnclosingLambda() != null) {
-                    tree.sym.owner = owner();
-                    if (tree.sym.hasOuterInstance()) {
-                        //if a class is defined within a lambda, the lambda must capture
-                        //its enclosing instance (if any)
-                        TranslationContext<?> localContext = context();
-                        final TypeSymbol outerInstanceSymbol = tree.sym.type.getEnclosingType().tsym;
-                        while (localContext != null && !localContext.owner.isStatic()) {
-                            if (localContext.tree.hasTag(LAMBDA)) {
-                                JCTree block = capturedDecl(localContext.depth, outerInstanceSymbol);
-                                if (block == null) break;
-                                ((LambdaTranslationContext)localContext)
-                                        .addSymbol(outerInstanceSymbol, CAPTURED_THIS);
-                            }
-                            localContext = localContext.prev;
-                        }
-                    }
-                }
-                frameStack = frameStack.prepend(new Frame(tree));
-                super.visitClassDef(tree);
-            }
-            finally {
-                log.useSource(prevSource.getFile());
-                frameStack = prevStack;
-                lambdaCount = prevLambdaCount;
-                syntheticMethodNameCounts = prevSyntheticMethodNameCounts;
-                clinits = prevClinits;
-            }
-        }
-
-        @Override
-        public void visitIdent(JCIdent tree) {
-            if (context() != null && lambdaIdentSymbolFilter(tree.sym)) {
-                if (tree.sym.kind == VAR &&
-                        tree.sym.owner.kind == MTH &&
-                        tree.type.constValue() == null) {
-                    TranslationContext<?> localContext = context();
-                    while (localContext != null) {
-                        if (localContext.tree.getTag() == LAMBDA) {
-                            JCTree block = capturedDecl(localContext.depth, tree.sym);
-                            if (block == null) break;
-                            ((LambdaTranslationContext)localContext)
-                                    .addSymbol(tree.sym, CAPTURED_VAR);
-                        }
-                        localContext = localContext.prev;
-                    }
-                } else if (tree.sym.owner.kind == TYP) {
-                    TranslationContext<?> localContext = context();
-                    while (localContext != null  && !localContext.owner.isStatic()) {
-                        if (localContext.tree.hasTag(LAMBDA)) {
-                            JCTree block = capturedDecl(localContext.depth, tree.sym);
-                            if (block == null) break;
-                            switch (block.getTag()) {
-                                case CLASSDEF:
-                                    JCClassDecl cdecl = (JCClassDecl)block;
-                                    ((LambdaTranslationContext)localContext)
-                                            .addSymbol(cdecl.sym, CAPTURED_THIS);
-                                    break;
-                                default:
-                                    Assert.error("bad block kind");
-                            }
-                        }
-                        localContext = localContext.prev;
-                    }
-                }
-            }
-            super.visitIdent(tree);
-        }
-
-        @Override
-        public void visitLambda(JCLambda tree) {
-            analyzeLambda(tree, "lambda.stat");
-        }
-
-        private void analyzeLambda(JCLambda tree, JCExpression methodReferenceReceiver) {
-            // Translation of the receiver expression must occur first
-            JCExpression rcvr = translate(methodReferenceReceiver);
-            LambdaTranslationContext context = analyzeLambda(tree, "mref.stat.1");
-            if (rcvr != null) {
-                context.methodReferenceReceiver = rcvr;
-            }
-        }
-
-        private LambdaTranslationContext analyzeLambda(JCLambda tree, String statKey) {
-            List<Frame> prevStack = frameStack;
-            try {
-                LambdaTranslationContext context = new LambdaTranslationContext(tree);
-                frameStack = frameStack.prepend(new Frame(tree));
-                for (JCVariableDecl param : tree.params) {
-                    context.addSymbol(param.sym, PARAM);
-                    frameStack.head.addLocal(param.sym);
-                }
-                contextMap.put(tree, context);
-                super.visitLambda(tree);
-                context.complete();
-                if (dumpLambdaToMethodStats) {
-                    log.note(tree, diags.noteKey(statKey, context.needsAltMetafactory(), context.translatedSym));
-                }
-                return context;
-            }
-            finally {
-                frameStack = prevStack;
-            }
-        }
-
-        @Override
-        public void visitMethodDef(JCMethodDecl tree) {
-            List<Frame> prevStack = frameStack;
-            try {
-                frameStack = frameStack.prepend(new Frame(tree));
-                super.visitMethodDef(tree);
-            }
-            finally {
-                frameStack = prevStack;
-            }
-        }
-
-        @Override
-        public void visitNewClass(JCNewClass tree) {
-            TypeSymbol def = tree.type.tsym;
-            boolean inReferencedClass = currentlyInClass(def);
-            boolean isLocal = def.isDirectlyOrIndirectlyLocal();
-            if ((inReferencedClass && isLocal || lambdaNewClassFilter(context(), tree))) {
-                TranslationContext<?> localContext = context();
-                final TypeSymbol outerInstanceSymbol = tree.type.getEnclosingType().tsym;
-                while (localContext != null  && !localContext.owner.isStatic()) {
-                    if (localContext.tree.hasTag(LAMBDA)) {
-                        if (outerInstanceSymbol != null) {
-                            JCTree block = capturedDecl(localContext.depth, outerInstanceSymbol);
-                            if (block == null) break;
-                        }
-                        ((LambdaTranslationContext)localContext)
-                                .addSymbol(outerInstanceSymbol, CAPTURED_THIS);
-                    }
-                    localContext = localContext.prev;
-                }
-            }
-            super.visitNewClass(tree);
-            if (context() != null && !inReferencedClass && isLocal) {
-                LambdaTranslationContext lambdaContext = (LambdaTranslationContext)context();
-                captureLocalClassDefs(def, lambdaContext);
-            }
-        }
-        //where
-        void captureLocalClassDefs(Symbol csym, final LambdaTranslationContext lambdaContext) {
-//            JCClassDecl localCDef = localClassDefs.get(csym);
-//            if (localCDef != null && lambdaContext.freeVarProcessedLocalClasses.add(csym)) {
-//                BasicFreeVarCollector fvc = lower.new BasicFreeVarCollector() {
-//                    @Override
-//                    void addFreeVars(ClassSymbol c) {
-//                        captureLocalClassDefs(c, lambdaContext);
-//                    }
-//                    @Override
-//                    void visitSymbol(Symbol sym) {
-//                        if (sym.kind == VAR &&
-//                                sym.owner.kind == MTH &&
-//                                ((VarSymbol)sym).getConstValue() == null) {
-//                            TranslationContext<?> localContext = context();
-//                            while (localContext != null) {
-//                                if (localContext.tree.getTag() == LAMBDA) {
-//                                    JCTree block = capturedDecl(localContext.depth, sym);
-//                                    if (block == null) break;
-//                                    ((LambdaTranslationContext)localContext).addSymbol(sym, CAPTURED_VAR);
-//                                }
-//                                localContext = localContext.prev;
-//                            }
-//                        }
-//                    }
-//                };
-//                fvc.scan(localCDef);
-//            }
-        }
-        //where
-        boolean currentlyInClass(Symbol csym) {
-            for (Frame frame : frameStack) {
-                if (frame.tree.hasTag(JCTree.Tag.CLASSDEF)) {
-                    JCClassDecl cdef = (JCClassDecl) frame.tree;
-                    if (cdef.sym == csym) {
-                        return true;
-                    }
-                }
-            }
-            return false;
-        }
-
-        /**
-         * Method references to local class constructors, may, if the local
-         * class references local variables, have implicit constructor
-         * parameters added in Lower; As a result, the invokedynamic bootstrap
-         * information added in the LambdaToMethod pass will have the wrong
-         * signature. Hooks between Lower and LambdaToMethod have been added to
-         * handle normal "new" in this case. This visitor converts potentially
-         * affected method references into a lambda containing a normal
-         * expression.
-         *
-         * @param tree
-         */
-        @Override
-        public void visitReference(JCMemberReference tree) {
-            ReferenceTranslationContext rcontext = new ReferenceTranslationContext(tree);
-            contextMap.put(tree, rcontext);
-            if (rcontext.needsConversionToLambda()) {
-                // Convert to a lambda, and process as such
-                MemberReferenceToLambda conv = new MemberReferenceToLambda(tree, rcontext, owner());
-                analyzeLambda(conv.lambda(), conv.getReceiverExpression());
-            } else {
-                super.visitReference(tree);
-                if (dumpLambdaToMethodStats) {
-                    // TODO
-//                    log.note(tree, Notes.MrefStat(rcontext.needsAltMetafactory(), null));
-                }
-            }
-        }
-
-        @Override
-        public void visitSelect(JCFieldAccess tree) {
-            if (context() != null && tree.sym.kind == VAR &&
-                    (tree.sym.name == names._this ||
-                            tree.sym.name == names._super)) {
-                // A select of this or super means, if we are in a lambda,
-                // we much have an instance context
-                TranslationContext<?> localContext = context();
-                while (localContext != null  && !localContext.owner.isStatic()) {
-                    if (localContext.tree.hasTag(LAMBDA)) {
-                        JCClassDecl clazz = (JCClassDecl)capturedDecl(localContext.depth, tree.sym);
-                        if (clazz == null) break;
-                        ((LambdaTranslationContext)localContext).addSymbol(clazz.sym, CAPTURED_THIS);
-                    }
-                    localContext = localContext.prev;
-                }
-            }
-            super.visitSelect(tree);
-        }
-
-        @Override
-        public void visitVarDef(JCVariableDecl tree) {
-            TranslationContext<?> context = context();
-            if (context != null && context instanceof LambdaTranslationContext lambdaContext) {
-                for (Frame frame : frameStack) {
-                    if (frame.tree.hasTag(VARDEF)) {
-                        //skip variable frames inside a lambda:
-                        continue;
-                    } else if (frame.tree.hasTag(LAMBDA)) {
-                        lambdaContext.addSymbol(tree.sym, LOCAL_VAR);
-                    } else {
-                        break;
-                    }
-                }
-                // Check for type variables (including as type arguments).
-                // If they occur within class nested in a lambda, mark for erasure
-                Type type = tree.sym.asType();
-            }
-
-            List<Frame> prevStack = frameStack;
-            try {
-                if (tree.sym.owner.kind == MTH) {
-                    frameStack.head.addLocal(tree.sym);
-                }
-                frameStack = frameStack.prepend(new Frame(tree));
-                super.visitVarDef(tree);
-            }
-            finally {
-                frameStack = prevStack;
-            }
-        }
-
-        /**
-         * Return a valid owner given the current declaration stack
-         * (required to skip synthetic lambda symbols)
-         */
-        private Symbol owner() {
-            return owner(false);
-        }
-
-        @SuppressWarnings("fallthrough")
-        private Symbol owner(boolean skipLambda) {
-            List<Frame> frameStack2 = frameStack;
-            while (frameStack2.nonEmpty()) {
-                switch (frameStack2.head.tree.getTag()) {
-                    case VARDEF:
-                        if (((JCVariableDecl)frameStack2.head.tree).sym.isDirectlyOrIndirectlyLocal()) {
-                            frameStack2 = frameStack2.tail;
-                            break;
-                        }
-                        JCClassDecl cdecl = (JCClassDecl)frameStack2.tail.head.tree;
-                        return initSym(cdecl.sym,
-                                ((JCVariableDecl)frameStack2.head.tree).sym.flags() & STATIC);
-                    case BLOCK:
-                        JCClassDecl cdecl2 = (JCClassDecl)frameStack2.tail.head.tree;
-                        return initSym(cdecl2.sym,
-                                ((JCBlock)frameStack2.head.tree).flags & STATIC);
-                    case CLASSDEF:
-                        return ((JCClassDecl)frameStack2.head.tree).sym;
-                    case METHODDEF:
-                        return ((JCMethodDecl)frameStack2.head.tree).sym;
-                    case LAMBDA:
-                        if (!skipLambda)
-                            return ((LambdaTranslationContext)contextMap
-                                    .get(frameStack2.head.tree)).translatedSym;
-                    default:
-                        frameStack2 = frameStack2.tail;
-                }
-            }
-            Assert.error();
-            return null;
-        }
-
-        private Symbol initSym(ClassSymbol csym, long flags) {
-            boolean isStatic = (flags & STATIC) != 0;
-            if (isStatic) {
-                /* static clinits are generated in Gen, so we need to use a fake
-                 * one. Attr creates a fake clinit method while attributing
-                 * lambda expressions used as initializers of static fields, so
-                 * let's use that one.
-                 */
-                MethodSymbol clinit = attr.removeClinit(csym);
-                if (clinit != null) {
-                    clinits.put(csym, clinit);
-                    return clinit;
-                }
-
-                /* if no clinit is found at Attr, then let's try at clinits.
-                 */
-                clinit = (MethodSymbol)clinits.get(csym);
-                if (clinit == null) {
-                    /* no luck, let's create a new one
-                     */
-                    clinit = makePrivateSyntheticMethod(STATIC,
-                            names.clinit,
-                            new MethodType(List.nil(), syms.voidType,
-                                    List.nil(), syms.methodClass),
-                            csym);
-                    clinits.put(csym, clinit);
-                }
-                return clinit;
-            } else {
-                //get the first constructor and treat it as the instance init sym
-                for (Symbol s : csym.members_field.getSymbolsByName(names.init)) {
-                    return s;
-                }
-            }
-            Assert.error("init not found");
-            return null;
-        }
-
-        private JCTree directlyEnclosingLambda() {
-            if (frameStack.isEmpty()) {
-                return null;
-            }
-            List<Frame> frameStack2 = frameStack;
-            while (frameStack2.nonEmpty()) {
-                switch (frameStack2.head.tree.getTag()) {
-                    case CLASSDEF:
-                    case METHODDEF:
-                        return null;
-                    case LAMBDA:
-                        return frameStack2.head.tree;
-                    default:
-                        frameStack2 = frameStack2.tail;
-                }
-            }
-            Assert.error();
-            return null;
-        }
-
-        private boolean inClassWithinLambda() {
-            if (frameStack.isEmpty()) {
-                return false;
-            }
-            List<Frame> frameStack2 = frameStack;
-            boolean classFound = false;
-            while (frameStack2.nonEmpty()) {
-                switch (frameStack2.head.tree.getTag()) {
-                    case LAMBDA:
-                        return classFound;
-                    case CLASSDEF:
-                        classFound = true;
-                        frameStack2 = frameStack2.tail;
-                        break;
-                    default:
-                        frameStack2 = frameStack2.tail;
-                }
-            }
-            // No lambda
-            return false;
-        }
-
-        /**
-         * Return the declaration corresponding to a symbol in the enclosing
-         * scope; the depth parameter is used to filter out symbols defined
-         * in nested scopes (which do not need to undergo capture).
-         */
-        private JCTree capturedDecl(int depth, Symbol sym) {
-            int currentDepth = frameStack.size() - 1;
-            for (Frame block : frameStack) {
-                switch (block.tree.getTag()) {
-                    case CLASSDEF:
-                        ClassSymbol clazz = ((JCClassDecl)block.tree).sym;
-                        if (clazz.isSubClass(sym, types) || sym.isMemberOf(clazz, types)) {
-                            return currentDepth > depth ? null : block.tree;
-                        }
-                        break;
-                    case VARDEF:
-                        if ((((JCVariableDecl)block.tree).sym == sym &&
-                                sym.owner.kind == MTH) || //only locals are captured
-                                (block.locals != null && block.locals.contains(sym))) {
-                            return currentDepth > depth ? null : block.tree;
-                        }
-                        break;
-                    case BLOCK:
-                    case METHODDEF:
-                    case LAMBDA:
-                        if (block.locals != null && block.locals.contains(sym)) {
-                            return currentDepth > depth ? null : block.tree;
-                        }
-                        break;
-                    default:
-                        Assert.error("bad decl kind " + block.tree.getTag());
-                }
-                currentDepth--;
-            }
-            return null;
-        }
-
-        private TranslationContext<?> context() {
-            for (Frame frame : frameStack) {
-                TranslationContext<?> context = contextMap.get(frame.tree);
-                if (context != null) {
-                    return context;
-                }
-            }
-            return null;
-        }
-
-        /**
-         *  This is used to filter out those identifiers that needs to be adjusted
-         *  when translating away lambda expressions
-         */
-        private boolean lambdaIdentSymbolFilter(Symbol sym) {
-            return (sym.kind == VAR || sym.kind == MTH)
-                    && !sym.isStatic()
-                    && sym.name != names.init;
-        }
-
-        /**
-         *  This is used to filter out those select nodes that need to be adjusted
-         *  when translating away lambda expressions - at the moment, this is the
-         *  set of nodes that select `this' (qualified this)
-         */
-        private boolean lambdaFieldAccessFilter(JCFieldAccess fAccess) {
-            return (context instanceof LambdaTranslationContext lambdaContext)
-                    && !fAccess.sym.isStatic()
-                    && fAccess.name == names._this
-                    && (fAccess.sym.owner.kind == TYP)
-                    && !lambdaContext.translatedSymbols.get(CAPTURED_OUTER_THIS).isEmpty();
-        }
-
-        /**
-         * This is used to filter out those new class expressions that need to
-         * be qualified with an enclosing tree
-         */
-        private boolean lambdaNewClassFilter(TranslationContext<?> context, JCNewClass tree) {
-            if (context != null
-                    && tree.encl == null
-                    && tree.def == null
-                    && !tree.type.getEnclosingType().hasTag(NONE)) {
-                Type encl = tree.type.getEnclosingType();
-                Type current = context.owner.enclClass().type;
-                while (!current.hasTag(NONE)) {
-                    if (current.tsym.isSubClass(encl.tsym, types)) {
-                        return true;
-                    }
-                    current = current.getEnclosingType();
-                }
-                return false;
-            } else {
-                return false;
-            }
-        }
-
-        private class Frame {
-            final JCTree tree;
-            List<Symbol> locals;
-
-            public Frame(JCTree tree) {
-                this.tree = tree;
-            }
-
-            void addLocal(Symbol sym) {
-                if (locals == null) {
-                    locals = List.nil();
-                }
-                locals = locals.prepend(sym);
-            }
-        }
-
-        /**
-         * This class is used to store important information regarding translation of
-         * lambda expression/method references (see subclasses).
-         */
-        abstract class TranslationContext<T extends JCFunctionalExpression> {
-
-            /** the underlying (untranslated) tree */
-            final T tree;
-
-            /** points to the adjusted enclosing scope in which this lambda/mref expression occurs */
-            final Symbol owner;
-
-            /** the depth of this lambda expression in the frame stack */
-            final int depth;
-
-            /** the enclosing translation context (set for nested lambdas/mref) */
-            final TranslationContext<?> prev;
-
-            /** list of methods to be bridged by the meta-factory */
-            final List<Symbol> bridges;
-
-            TranslationContext(T tree) {
-                this.tree = tree;
-                this.owner = owner(true);
-                this.depth = frameStack.size() - 1;
-                this.prev = context();
-                ClassSymbol csym =
-                        types.makeFunctionalInterfaceClass(attrEnv, names.empty, tree.target, ABSTRACT | INTERFACE);
-                this.bridges = types.functionalInterfaceBridges(csym);
-            }
-
-            /** does this functional expression need to be created using alternate metafactory? */
-            boolean needsAltMetafactory() {
-                return tree.target.isIntersection() ||
-                        bridges.length() > 1;
-            }
-
-            /**
-             * @return Name of the enclosing method to be folded into synthetic
-             * method name
-             */
-            String enclosingMethodName() {
-                return syntheticMethodNameComponent(owner.name);
-            }
-
-            /**
-             * @return Method name in a form that can be folded into a
-             * component of a synthetic method name
-             */
-            String syntheticMethodNameComponent(Name name) {
-                if (name == null) {
-                    return "null";
-                }
-                String methodName = name.toString();
-                if (methodName.equals("<clinit>")) {
-                    methodName = "static";
-                } else if (methodName.equals("<init>")) {
-                    methodName = "new";
-                }
-                return methodName;
-            }
-        }
-
-        /**
-         * This class retains all the useful information about a lambda expression;
-         * the contents of this class are filled by the LambdaAnalyzer visitor,
-         * and the used by the main translation routines in order to adjust references
-         * to captured locals/members, etc.
-         */
-        class LambdaTranslationContext extends TranslationContext<JCLambda> {
-
-            /** variable in the enclosing context to which this lambda is assigned */
-            final Symbol self;
-
-            /** variable in the enclosing context to which this lambda is assigned */
-            final Symbol assignedTo;
-
-            Map<LambdaSymbolKind, Map<Symbol, Symbol>> translatedSymbols;
-
-            /** the synthetic symbol for the method hoisting the translated lambda */
-            MethodSymbol translatedSym;
-
-            List<JCVariableDecl> syntheticParams;
-
-            /**
-             * to prevent recursion, track local classes processed
-             */
-            final Set<Symbol> freeVarProcessedLocalClasses;
-
-            /**
-             * For method references converted to lambdas.  The method
-             * reference receiver expression. Must be treated like a captured
-             * variable.
-             */
-            JCExpression methodReferenceReceiver;
-
-            LambdaTranslationContext(JCLambda tree) {
-                super(tree);
-                Frame frame = frameStack.head;
-                switch (frame.tree.getTag()) {
-                    case VARDEF:
-                        assignedTo = self = ((JCVariableDecl) frame.tree).sym;
-                        break;
-                    case ASSIGN:
-                        self = null;
-                        assignedTo = TreeInfo.symbol(((JCAssign) frame.tree).getVariable());
-                        break;
-                    default:
-                        assignedTo = self = null;
-                        break;
-                }
-
-                // This symbol will be filled-in in complete
-                if (owner.kind == MTH) {
-                    final MethodSymbol originalOwner = (MethodSymbol)owner.clone(owner.owner);
-                    this.translatedSym = new MethodSymbol(SYNTHETIC | PRIVATE, null, null, owner.enclClass()) {
-                        @Override
-                        public MethodSymbol originalEnclosingMethod() {
-                            return originalOwner;
-                        }
-                    };
-                } else {
-                    this.translatedSym = makePrivateSyntheticMethod(0, null, null, owner.enclClass());
-                }
-                translatedSymbols = new EnumMap<>(LambdaSymbolKind.class);
-
-                translatedSymbols.put(PARAM, new LinkedHashMap<Symbol, Symbol>());
-                translatedSymbols.put(LOCAL_VAR, new LinkedHashMap<Symbol, Symbol>());
-                translatedSymbols.put(CAPTURED_VAR, new LinkedHashMap<Symbol, Symbol>());
-                translatedSymbols.put(CAPTURED_THIS, new LinkedHashMap<Symbol, Symbol>());
-                translatedSymbols.put(CAPTURED_OUTER_THIS, new LinkedHashMap<Symbol, Symbol>());
-
-                freeVarProcessedLocalClasses = new HashSet<>();
-            }
-
-            /**
-             * For a non-serializable lambda, generate a simple method.
-             *
-             * @return Name to use for the synthetic lambda method name
-             */
-            private Name lambdaName() {
-                return names.lambda.append(names.fromString(enclosingMethodName() + "$" + lambdaCount++));
-            }
-
-            /**
-             * Translate a symbol of a given kind into something suitable for the
-             * synthetic lambda body
-             */
-            Symbol translate(final Symbol sym, LambdaSymbolKind skind) {
-                Symbol ret;
-                switch (skind) {
-                    case CAPTURED_THIS:
-                        ret = sym;  // self represented
-                        break;
-                    case CAPTURED_VAR:
-                        ret = new VarSymbol(SYNTHETIC | FINAL | PARAMETER, sym.name, sym.type, translatedSym) {
-                            @Override
-                            public Symbol baseSymbol() {
-                                //keep mapping with original captured symbol
-                                return sym;
-                            }
-                        };
-                        break;
-                    case CAPTURED_OUTER_THIS:
-                        Name name = names.fromString(sym.flatName().toString().replace('.', '$') + names.dollarThis);
-                        ret = new VarSymbol(SYNTHETIC | FINAL | PARAMETER, name, sym.type, translatedSym) {
-                            @Override
-                            public Symbol baseSymbol() {
-                                //keep mapping with original captured symbol
-                                return sym;
-                            }
-                        };
-                        break;
-                    case LOCAL_VAR:
-                        ret = new VarSymbol(sym.flags() & FINAL, sym.name, sym.type, translatedSym) {
-                            @Override
-                            public Symbol baseSymbol() {
-                                //keep mapping with original symbol
-                                return sym;
-                            }
-                        };
-                        ((VarSymbol) ret).pos = ((VarSymbol) sym).pos;
-                        // If sym.data == ElementKind.EXCEPTION_PARAMETER,
-                        // set ret.data = ElementKind.EXCEPTION_PARAMETER too.
-                        // Because method com.sun.tools.javac.jvm.Code.fillExceptionParameterPositions and
-                        // com.sun.tools.javac.jvm.Code.fillLocalVarPosition would use it.
-                        // See JDK-8257740 for more information.
-                        if (((VarSymbol) sym).isExceptionParameter()) {
-                            ((VarSymbol) ret).setData(ElementKind.EXCEPTION_PARAMETER);
-                        }
-                        break;
-                    case PARAM:
-                        ret = new VarSymbol((sym.flags() & FINAL) | PARAMETER, sym.name, sym.type, translatedSym);
-                        ((VarSymbol) ret).pos = ((VarSymbol) sym).pos;
-                        // Set ret.data. Same as case LOCAL_VAR above.
-                        if (((VarSymbol) sym).isExceptionParameter()) {
-                            ((VarSymbol) ret).setData(ElementKind.EXCEPTION_PARAMETER);
-                        }
-                        break;
-                    default:
-                        Assert.error(skind.name());
-                        throw new AssertionError();
-                }
-                if (ret != sym && skind.propagateAnnotations()) {
-                    ret.setDeclarationAttributes(sym.getRawAttributes());
-                    ret.setTypeAttributes(sym.getRawTypeAttributes());
-                }
-                return ret;
-            }
-
-            void addSymbol(Symbol sym, LambdaSymbolKind skind) {
-                if (skind == CAPTURED_THIS && sym != null && sym.kind == TYP && !typesUnderConstruction.isEmpty()) {
-                    ClassSymbol currentClass = currentClass();
-                    if (currentClass != null && typesUnderConstruction.contains(currentClass)) {
-                        // reference must be to enclosing outer instance, mutate capture kind.
-                        Assert.check(sym != currentClass); // should have been caught right in Attr
-                        skind = CAPTURED_OUTER_THIS;
-                    }
-                }
-                Map<Symbol, Symbol> transMap = getSymbolMap(skind);
-                if (!transMap.containsKey(sym)) {
-                    transMap.put(sym, translate(sym, skind));
-                }
-            }
-
-            Map<Symbol, Symbol> getSymbolMap(LambdaSymbolKind skind) {
-                Map<Symbol, Symbol> m = translatedSymbols.get(skind);
-                Assert.checkNonNull(m);
-                return m;
-            }
-
-            JCTree translate(JCIdent lambdaIdent) {
-                for (LambdaSymbolKind kind : LambdaSymbolKind.values()) {
-                    Map<Symbol, Symbol> m = getSymbolMap(kind);
-                    switch(kind) {
-                        default:
-                            if (m.containsKey(lambdaIdent.sym)) {
-                                Symbol tSym = m.get(lambdaIdent.sym);
-                                JCTree t = makeIdent(tSym).setType(lambdaIdent.type);
-                                return t;
-                            }
-                            break;
-                        case CAPTURED_OUTER_THIS:
-                            Optional<Symbol> proxy = m.keySet().stream()
-                                    .filter(out -> lambdaIdent.sym.isMemberOf(out.type.tsym, types))
-                                    .reduce((a, b) -> a.isEnclosedBy((ClassSymbol)b) ? a : b);
-                            if (proxy.isPresent()) {
-                                // Transform outer instance variable references anchoring them to the captured synthetic.
-                                Symbol tSym = m.get(proxy.get());
-                                JCExpression t = makeIdent(tSym).setType(lambdaIdent.sym.owner.type);
-                                t = make.Select(t, lambdaIdent.sym);
-                                t.setType(lambdaIdent.type);
-                                TreeInfo.setSymbol(t, lambdaIdent.sym);
-                                return t;
-                            }
-                            break;
-                    }
-                }
-                return null;
-            }
-
-            /* Translate away qualified this expressions, anchoring them to synthetic parameters that
-               capture the qualified this handle. `fieldAccess' is guaranteed to one such.
-            */
-            public JCTree translate(JCFieldAccess fieldAccess) {
-                Assert.check(fieldAccess.name == names._this);
-                Map<Symbol, Symbol> m = translatedSymbols.get(LambdaSymbolKind.CAPTURED_OUTER_THIS);
-                if (m.containsKey(fieldAccess.sym.owner)) {
-                    Symbol tSym = m.get(fieldAccess.sym.owner);
-                    JCExpression t = makeIdent(tSym).setType(fieldAccess.sym.owner.type);
-                    return t;
-                }
-                return null;
-            }
-
-            /* Translate away naked new instance creation expressions with implicit enclosing instances,
-               anchoring them to synthetic parameters that stand proxy for the qualified outer this handle.
-            */
-            public JCNewClass translate(JCNewClass newClass) {
-                Assert.check(newClass.clazz.type.tsym.hasOuterInstance() && newClass.encl == null);
-                Map<Symbol, Symbol> m = translatedSymbols.get(LambdaSymbolKind.CAPTURED_OUTER_THIS);
-                final Type enclosingType = newClass.clazz.type.getEnclosingType();
-                if (m.containsKey(enclosingType.tsym)) {
-                    Symbol tSym = m.get(enclosingType.tsym);
-                    JCExpression encl = makeIdent(tSym).setType(enclosingType);
-                    newClass.encl = encl;
-                }
-                return newClass;
-            }
-
-            /**
-             * The translatedSym is not complete/accurate until the analysis is
-             * finished.  Once the analysis is finished, the translatedSym is
-             * "completed" -- updated with type information, access modifiers,
-             * and full parameter list.
-             */
-            void complete() {
-                if (syntheticParams != null) {
-                    return;
-                }
-                boolean inInterface = translatedSym.owner.isInterface();
-                boolean thisReferenced = !getSymbolMap(CAPTURED_THIS).isEmpty();
-
-                // If instance access isn't needed, make it static.
-                // Interface instance methods must be default methods.
-                // Lambda methods are private synthetic.
-                // Inherit ACC_STRICT from the enclosing method, or, for clinit,
-                // from the class.
-                translatedSym.flags_field = SYNTHETIC | LAMBDA_METHOD |
-                        owner.flags_field & STRICTFP |
-                        owner.owner.flags_field & STRICTFP |
-                        PRIVATE |
-                        (thisReferenced? (inInterface? DEFAULT : 0) : STATIC);
-
-                //compute synthetic params
-                ListBuffer<JCVariableDecl> params = new ListBuffer<>();
-                ListBuffer<VarSymbol> parameterSymbols = new ListBuffer<>();
-
-                // The signature of the method is augmented with the following
-                // synthetic parameters:
-                //
-                // 1) reference to enclosing contexts captured by the lambda expression
-                // 2) enclosing locals captured by the lambda expression
-                for (Symbol thisSym : getSymbolMap(CAPTURED_VAR).values()) {
-                    params.append(make.VarDef((VarSymbol) thisSym, null));
-                    parameterSymbols.append((VarSymbol) thisSym);
-                }
-                for (Symbol thisSym : getSymbolMap(CAPTURED_OUTER_THIS).values()) {
-                    params.append(make.VarDef((VarSymbol) thisSym, null));
-                    parameterSymbols.append((VarSymbol) thisSym);
-                }
-                for (Symbol thisSym : getSymbolMap(PARAM).values()) {
-                    params.append(make.VarDef((VarSymbol) thisSym, null));
-                    parameterSymbols.append((VarSymbol) thisSym);
-                }
-                syntheticParams = params.toList();
-
-                translatedSym.params = parameterSymbols.toList();
-
-                // Compute and set the lambda name
-                translatedSym.name = lambdaName();
-
-                //prepend synthetic args to translated lambda method signature
-                translatedSym.type = types.createMethodTypeWithParameters(
-                        generatedLambdaSig(),
-                        TreeInfo.types(syntheticParams));
-            }
-
-            Type generatedLambdaSig() {
-                return tree.getDescriptorType(types);
-            }
-        }
-
-        /**
-         * This class retains all the useful information about a method reference;
-         * the contents of this class are filled by the LambdaAnalyzer visitor,
-         * and the used by the main translation routines in order to adjust method
-         * references (i.e. in case a bridge is needed)
-         */
-        final class ReferenceTranslationContext extends TranslationContext<JCMemberReference> {
-
-            final boolean isSuper;
-
-            ReferenceTranslationContext(JCMemberReference tree) {
-                super(tree);
-                this.isSuper = tree.hasKind(ReferenceKind.SUPER);
-            }
-
-            boolean needsVarArgsConversion() {
-                return tree.varargsElement != null;
-            }
-
-            /**
-             * @return Is this an array operation like clone()
-             */
-            boolean isArrayOp() {
-                return tree.sym.owner == syms.arrayClass;
-            }
-
-            boolean receiverAccessible() {
-                //hack needed to workaround 292 bug (7087658)
-                //when 292 issue is fixed we should remove this and change the backend
-                //code to always generate a method handle to an accessible method
-                return tree.ownerAccessible;
-            }
-
-            /**
-             * This method should be called only when target release <= 14
-             * where LambdaMetaFactory does not spin nestmate classes.
-             *
-             * This method should be removed when --release 14 is not supported.
-             */
-            boolean isPrivateInOtherClass() {
-                assert !nestmateLambdas;
-                return  (tree.sym.flags() & PRIVATE) != 0 &&
-                        !types.isSameType(
-                                tree.sym.enclClass().asType(),
-                                owner.enclClass().asType());
-            }
-
-            /**
-             * Erasure destroys the implementation parameter subtype
-             * relationship for intersection types.
-             * Have similar problems for union types too.
-             */
-            boolean interfaceParameterIsIntersectionOrUnionType() {
-                List<Type> tl = tree.getDescriptorType(types).getParameterTypes();
-                for (; tl.nonEmpty(); tl = tl.tail) {
-                    Type pt = tl.head;
-                    if (isIntersectionOrUnionType(pt))
-                        return true;
-                }
-                return false;
-            }
-
-            boolean isIntersectionOrUnionType(Type t) {
-                switch (t.getKind()) {
-                    case INTERSECTION:
-                    case UNION:
-                        return true;
-                    case TYPEVAR:
-                        TypeVar tv = (TypeVar) t;
-                        return isIntersectionOrUnionType(tv.getUpperBound());
-                }
-                return false;
-            }
-
-            /**
-             * Does this reference need to be converted to a lambda
-             * (i.e. var args need to be expanded or "super" is used)
-             */
-            final boolean needsConversionToLambda() {
-                return interfaceParameterIsIntersectionOrUnionType() ||
-                        isSuper ||
-                        needsVarArgsConversion() ||
-                        isArrayOp() ||
-                        (!nestmateLambdas && isPrivateInOtherClass()) ||
-                        isProtectedInSuperClassOfEnclosingClassInOtherPackage(tree.sym, owner) ||
-                        !receiverAccessible() ||
-                        (tree.getMode() == ReferenceMode.NEW &&
-                                tree.kind != ReferenceKind.ARRAY_CTOR &&
-                                (tree.sym.owner.isDirectlyOrIndirectlyLocal() || tree.sym.owner.isInner()));
-            }
-
-            Type generatedRefSig() {
-                return tree.sym.type;
-            }
-
-            Type bridgedRefSig() {
-                return types.findDescriptorSymbol(tree.target.tsym).type;
-            }
-        }
-    }
-
-    private JCIdent makeIdent(Symbol tSym) {
+    public JCIdent makeIdent(Symbol tSym) {
         var result = make.Ident(tSym);
         result.sym = tSym;
         return result;
@@ -1865,8 +590,8 @@ public class LambdaCompiler extends TreeTranslator {
         }
     }
 
-    private boolean isProtectedInSuperClassOfEnclosingClassInOtherPackage(Symbol targetReference,
-                                                                          Symbol currentClass) {
+    public boolean isProtectedInSuperClassOfEnclosingClassInOtherPackage(Symbol targetReference,
+                                                                         Symbol currentClass) {
         return ((targetReference.flags() & PROTECTED) != 0 &&
                 targetReference.packge() != currentClass.packge());
     }
